Zavrseni svi api endpointovi i testirano

+++ app/Http/Controllers/Api/ReservationController.php
@@ -4,29 +4,48 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Reservation;
+use Carbon\Carbon;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Database\Eloquent\Collection;
+use App\Rules\MoreThan24H;
 
 class ReservationController extends Controller
 {
-    //
-    public function index(): Collection
+    public function index(): JsonResponse
     {
-        return Reservation::all();
+        return response()->json(Reservation::all());
     }
 
     public function store(Request $request): JsonResponse
     {
-        $reservation = new Reservation($request->validate([
-            'start_date' => 'required|date|after:now',
+        $request->validate([
+            'start_date' => ['required', 'date', new MoreThan24H()],
             'end_date' => 'required|date|after:start_date',
             'user_id' => 'required|exists:users,id',
-            'car_id' => 'required|exists:cars,id',
-            'price' => 'required|numeric'
-        ]));
+            'car_id' => 'required|exists:cars,id'
+        ]);
+
+        $reservation = $this->createReservation($request);
         $reservation->save();
-        return response()->json($reservation);
+        return response()->json($reservation->withoutRelations());
+    }
+
+    public function createReservation(Request $request): Reservation
+    {
+        $reservation = new Reservation();
+        $reservation->start_date = $request->start_date;
+        $reservation->end_date = $request->end_date;
+        $reservation->user_id = $request->user_id;
+        $reservation->car_id = $request->car_id;
+        $reservation->price = $this->calculatePrice($reservation);
+        return $reservation;
+    }
+
+    public function calculatePrice(Reservation $reservation): float
+    {
+        $days = ceil(Carbon::parse($reservation->start_date)
+            ->diffInDays(Carbon::parse($reservation->end_date)));
+        return $days * $reservation->car->price;
     }
 
     public function show(Reservation $reservation): JsonResponse
@@ -37,19 +56,25 @@ class ReservationController extends Controller
     public function update(Request $request, Reservation $reservation): JsonResponse
     {
         $reservation->update($request->validate([
-            'start_date' => 'sometimes|date|after:now',
-            'end_date' => 'sometimes|date|after:start_date',
-            'user_id' => 'sometimes|exists:users,id',
-            'car_id' => 'sometimes|exists:cars,id',
-            'price' => 'sometimes|numeric'
+            'start_date' => 'required|date|after:now',
+            'end_date' => 'required|date|after:start_date',
+            'user_id' => 'required|exists:users,id',
+            'car_id' => 'required|exists:cars,id'
         ]));
 
-        return response()->json($reservation);
+        $reservation->price = $this->calculatePrice($reservation);
+        $reservation->save();
+        return response()->json($reservation->withoutRelations());
     }
 
     public function destroy(Reservation $reservation): JsonResponse
     {
-        $reservation->delete();
-        return response()->json(null, 204);
+        if($reservation->start_date >= now()->addHours(48)) {
+            $reservation->delete();
+            return response()->json(['message' => 'Reservation canceled'], 204);
+        }
+        else{
+            return response()->json('Reservations can be cancelled only 48 hours prior start without charge' , 422);
+        }
     }
 }
